requirements.txt
Flask==3.1.1
python-dotenv==1.1.0
Flask-SocketIO==5.3.6
redis==5.0.0
mysql-connector-python==8.1.0
python-socketio==5.8.0
APScheduler==3.10.4
flask-cors